refactor: Wrap APIs call in NetworkManager
