+++ PokemonViewer/PokemonTableViewCell.swift
@@ -11,8 +11,8 @@ class PokemonTableViewCell: UITableViewCell {
     let thumbnailImageView = UIImageView()
     let nameLabel = UILabel()
     var cellShouldUpdate: (() -> Void)?
-    private let imageCache = NSCache<NSURL, UIImage>()
     private var imageFetchTask: Task<Void, Never>?
+    private let networkManager: NetworkManager = .shared
     
     override init(style: UITableViewCell.CellStyle, reuseIdentifier: String?) {
         super.init(style: style, reuseIdentifier: reuseIdentifier)
@@ -42,33 +42,9 @@ class PokemonTableViewCell: UITableViewCell {
     func setupContent(_ name: String, _ imageUrl: URL?) {
         self.nameLabel.text = name
         imageFetchTask = Task {
-            self.thumbnailImageView.image = await self.fetchImage(from: imageUrl)
+            self.thumbnailImageView.image = await networkManager.fetchImage(from: imageUrl)
+            self.cellShouldUpdate?()
         }
-    }
-    
-    func fetchImage(from url: URL?) async -> UIImage? {
-        guard let url = url else { return nil }
-        
-        if let cachedImage = imageCache.object(forKey: url as NSURL) {
-            return cachedImage
-        }
-        
-        do {
-            if let image = imageCache.object(forKey: url as NSURL) {
-                self.cellShouldUpdate?()
-                return image
-            } else {
-                let (data, _) = try await URLSession.shared.data(from: url)
-                guard let image = UIImage(data: data) else { return nil }
-                self.imageCache.setObject(image, forKey: url as NSURL)
-                self.cellShouldUpdate?()
-                return image
-            }
-        } catch {
-            print(String(describing: error))
-        }
-        
-        return nil
     }
     
     override func prepareForReuse() {
